[TASK] view now in separate file, blackified code
